Update EngineIgnitor.cs
changed requirements for ignitions to work when ignitionsremained are negative numbers. Infinite ignition engines only work once because when the ignite, the remaining ignitions drop from -1 (infinite) to -2 (?¿), meaning they act like they no longer have any ignitions avalible. 

This is of cource a crude way to fix it, since the correct way would be too add an exception for when the engines have infinite ignitions, but my code skills are not there yet.